Crate viper

Re-exports

pub use crate::silicon_counterexample::*;

Modules

errors

silicon_counterexample

smt_manager

utils

Various utility methods for working with Viper.

Macros

build_ast_node

build_ast_node_with_pos

generate_conversion_from_to

get_ast_object

jobject_wrapper

map_to_jobject

map_to_jobject_pair

map_to_jobject_pairs

map_to_jobjects

run_timed

Runs statements on the same level as the macro call, timing and logging (info-level by default) how long it took.

Structs

AstFactory

AstUtils

ConsistencyError

The consistency error reported by the verifier.

Declaration

Domain

DomainFunc

Expr

Field

Function

JavaException

The structure describing a Java exception

JavaExceptionWithOrigin

The Java exception reported by the verifier.

LocalVarDecl

Location

Method

NamedDomainAxiom

PersistentCache

Position

Predicate

Program

Stmt

Trigger

Type

UknownBackendError

VerificationContext

VerificationError

Verifier

Viper

Enums

BinOpBv

BinOpFloat

BvSize

FloatSizeViper

UnOpBv

UnOpFloat

VerificationBackend

VerificationResult

The result of a verification request on a Viper program.

Traits

Cache