Trait viper::utils::ExprIterator

pub trait ExprIterator<'v> {
    fn conjoin(&mut self, ast: &AstFactory<'v>) -> Expr<'v>;
    fn conjoin_with_init(
        &mut self, 
        ast: &AstFactory<'v>, 
        init: Expr<'v>
    ) -> Expr<'v>;
}

Required Methods

fn conjoin(&mut self, ast: &AstFactory<'v>) -> Expr<'v>

Conjoin a sequence of expressions into a single expression. Panics if the sequence has no elements.

fn conjoin_with_init(
&mut self,
ast: &AstFactory<'v>,
init: Expr<'v>
) -> Expr<'v>

Conjoin a sequence of expressions into a single expression.